requirements.txt
astroid==2.4.2
backcall==0.2.0
certifi==2020.6.20
cloudpickle==1.3.0
cma==3.0.3
cocoex==2.3.3.84
cocopp==2.3.3.18
cycler==0.10.0
decorator==4.4.2
-e git+https://github.com/optimization-toolbox/DNE4py@4639ff4f869f07d6f5858edf12bbc96e2c2c1824#egg=DNE4py
emcee==2.2.1
emukit==0.4.7
future==0.18.2
GPy==1.9.9
gym==0.17.3
-e git+https://github.com/optimization-toolbox/GYM4bbo.git@9d1b4ebf44d8f66d71cac391b35e4ffa658d5604#egg=GYM4bbo
IOHexperimenter==0.2.8.7
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.5.3
jedi==0.17.2
jupyter-client==6.1.7
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
llvmlite==0.34.0
matplotlib==3.1.0
mccabe==0.6.1
memory-profiler==0.58.0
-e git+https://github.com/Hugodovs/meta-learning-optimizers.git@be68d2f7bc6a3ecd4d0fbaf24e517592f3eef8fe#egg=meta_learning_optimizers
mpi4py==3.0.3
numba==0.51.2
numpy==1.19.1
paramz==0.9.5
parso==0.7.1
pexpect==4.8.0
-e git+https://github.com/optimization-toolbox/PI4py.git@7d6dbab1894ba506e4cb1b08abf675d44865b451#egg=PI4py
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.14.1
prettytable==0.7.2
prompt-toolkit==3.0.7
psutil==5.7.3
ptyprocess==0.6.0
pybnn==0.0.5
pyglet==1.5.0
Pygments==2.7.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
retrying==1.3.3
scipy==1.1.0
six==1.15.0
snakeviz==2.1.0
sobol-seq==0.2.0
toml==0.10.1
torch==1.6.0
torchvision==0.7.0
tornado==6.0.4
traitlets==5.0.4
typed-ast==1.4.1
wcwidth==0.2.5
wrapt==1.12.1
